Update participant count type from Long to Integer

+++ src/main/java/com/pickup/organizer/exception/game/InvalidCapacityException.java
@@ -2,7 +2,7 @@ package com.pickup.organizer.exception.game;
 
 public class InvalidCapacityException extends RuntimeException {
 
-    public InvalidCapacityException(Integer capacity, Long currentParticipants) {
+    public InvalidCapacityException(Integer capacity, Integer currentParticipants) {
         super("Cannot reduce game capacity to '" + capacity + "'."
             + " There are already '" + currentParticipants + "' players registered.");
     }
